Fix the folder size in two more places. Thanks to Jean-Christophe for noticing this.
CCBUG: 151155


svn path=/trunk/KDE/kdepim/; revision=751307

+++ kmfoldertree.cpp
@@ -1682,7 +1682,7 @@ void KMFolderTree::slotUpdateCounts(KMFolder * folder)
   }
   if ( isSizeActive() ) {
     if ( !fti->folder()->noContent()) {
-      int size = folder->storage()->folderSize();
+      qint64 size = folder->storage()->folderSize();
       if ( size != fti->folderSize() ) {
         fti->setFolderSize( size );
         repaint = true;
